+++ internal/provider/data_source_utils_yaml_merge.go
@@ -98,6 +98,9 @@ func (d *yamlMergeDataSource) Read(ctx context.Context, req datasource.ReadReque
 			)
 			return
 		}
+		if decoded == nil {
+			continue
+		}
 
 		resolved, err := resolveYamlTags(decoded)
 		if err != nil {